fix: make the macOS crate compile on GitHub Actions
Match objc2 0.6 init/retain, libproc constants, Send for IOProc pointers, and AppKit method signatures that only show up when compiling on Darwin.

+++ src/paths.rs
@@ -205,7 +205,7 @@ fn free_bytes_impl(path: &Path) -> Option<u64> {
     if err != 0 {
         return None;
     }
-    Some(st.f_bavail.saturating_mul(st.f_frsize as u64))
+    Some((st.f_bavail as u64).saturating_mul(st.f_frsize as u64))
 }
 
 #[cfg(not(any(windows, target_os = "macos")))]
